+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 ## [Unreleased]
 
+### Added
+- Lane types walking, biking and curb
+- Cross section preset "Urban two lanes with curb and walkway" with two driving
+  lanes separated by a broken line and border, curb and walking lanes on each side
+- Junction connecting roads for walking and biking lanes, the connecting road
+  takes over the lane type and the curb height of the lanes it connects
+
+### Fixed
+- Junction joint snapping only offered driving lanes so that walking and biking
+  lanes of the incoming roads could not be connected
+
 ## [0.35.0] - 2026-09-08
 
 ### Added